// src/components/flow/types.ts
export interface Choice {
  id: string;
  content: string;
}

export type QuestionNodeType = {
  id: string;
  topic: string;
  img?: string;
  choices: Choice[];
};

export type ResultNodeType = {
  id: string;
  result: string;
  message?: string;
  img?: string;
  url?: string;
};

export type NodeType = "qNode" | "rNode";